moved movie details api call into its own serverless function
